+++ src/components/Tab/index.js
@@ -6,6 +6,26 @@ const VuePlugin = {
   install(Vue) {
     installPlugin(Vue, Tab);
     installPlugin(Vue, TabController);
+
+    // const ZToast = Toast;
+    
+    // Vue.mixin(ZToast);
+
+    Vue.prototype.$zutre.newTabController = function(parent) {
+
+      let newInstance = new Vue({
+        ...TabController,
+        parent: parent,
+        propsData: {
+          items: parent.items,
+          active: parent.activeTab,
+          height: parent.height,
+        },
+      });
+
+      return newInstance;
+    };
+
   },
 };